service-workers/service-worker/navigation-redirect.https.html

<!DOCTYPE html>
<title>Service Worker: Navigation redirection</title>
<meta name="timeout" content="long">
<!-- empty variant tests document.location and intercepted URLs -->
<meta name="variant" content="">
<!-- client variant tests the Clients API (resultingClientId and Client.url) -->
<meta name="variant" content="?client">

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/get-host-info.sub.js"></script>
<script src="resources/test-helpers.sub.js"></script>
<body>
<script>
const host_info = get_host_info();

// This test registers three Service Workers at SCOPE1, SCOPE2 and
// OTHER_ORIGIN_SCOPE. And checks the redirected page's URL and the requests
// which are intercepted by Service Worker while loading redirect page.
const BASE_URL = host_info['HTTPS_ORIGIN'] + base_path();
const OTHER_BASE_URL = host_info['HTTPS_REMOTE_ORIGIN'] + base_path();

const SCOPE1 = BASE_URL + 'resources/navigation-redirect-scope1.py?';
const SCOPE2 = BASE_URL + 'resources/navigation-redirect-scope2.py?';
const OUT_SCOPE = BASE_URL + 'resources/navigation-redirect-out-scope.py?';
const SCRIPT = 'resources/redirect-worker.js';

const OTHER_ORIGIN_IFRAME_URL =
      OTHER_BASE_URL + 'resources/navigation-redirect-other-origin.html';
const OTHER_ORIGIN_SCOPE =
      OTHER_BASE_URL + 'resources/navigation-redirect-scope1.py?';
const OTHER_ORIGIN_OUT_SCOPE =
      OTHER_BASE_URL + 'resources/navigation-redirect-out-scope.py?';

let registrations;
let workers;
let other_origin_frame;
let message_resolvers = {};
let next_message_id = 0;

promise_test(async t  => {
  // In this frame we register a service worker at OTHER_ORIGIN_SCOPE.
  // And will use this frame to communicate with the worker.
  other_origin_frame = await with_iframe(OTHER_ORIGIN_IFRAME_URL);

  // Register same-origin service workers.
  registrations = await Promise.all([
      service_worker_unregister_and_register(t, SCRIPT, SCOPE1),
      service_worker_unregister_and_register(t, SCRIPT, SCOPE2)]);

  // Wait for all workers to activate.
  workers = registrations.map(get_effective_worker);
  return Promise.all([
      wait_for_state(t, workers[0], 'activated'),
      wait_for_state(t, workers[1], 'activated'),
      // This promise will resolve when |wait_for_worker_promise|
      // in OTHER_ORIGIN_IFRAME_URL resolves.
      send_to_iframe(other_origin_frame, {command: 'wait_for_worker'})]);
}, 'initialize global state');

function get_effective_worker(registration) {
  if (registration.active)
    return registration.active;
  if (registration.waiting)
    return registration.waiting;
  if (registration.installing)
    return registration.installing;
}

async function check_all_intercepted_urls(expected_urls) {
  const urls = [];
  urls.push(await get_intercepted_urls(workers[0]));
  urls.push(await get_intercepted_urls(workers[1]));
  // Gets the request URLs which are intercepted by OTHER_ORIGIN_SCOPE's
  // SW. This promise will resolve when get_intercepted_urls() in
  // OTHER_ORIGIN_IFRAME_URL resolves.
  urls.push(await send_to_iframe(other_origin_frame, {command: 'get_intercepted_urls'}));
  assert_object_equals(urls, expected_urls, 'Intercepted URLs should match.');
}

async function get_all_request_infos()  {
  const request_infos = [];
  request_infos.push(await get_request_infos(workers[0]));
  request_infos.push(await get_request_infos(workers[1]));
  request_infos.push(await send_to_iframe(other_origin_frame, {command: 'get_request_infos'}));
  return request_infos;
}

// Checks |clients| returned from a worker. Only the client named 'x' should
// be found. Returns true if a client was found.
//
// Note that 'x' may not be found by this worker, if it's cross-origin.
//
// |clients| is an object like:
// {x: {found: true, id: id1, url: url1}, b: {found: false}}
function check_clients(clients, expected_id, expected_url, worker_name) {
  let found = false;
  Object.keys(clients).forEach(key => {
    const info = clients[key];
    if (info.found) {
      // Only x should be found.
      assert_equals(key, 'x', `${worker_name} client key`);

      assert_equals(info.id, expected_id, `${worker_name} client id`);
      assert_equals(info.url, expected_url, `${worker_name} client url`);
      found = true;
    }
  });
  return found;
}

function check_resulting_client_ids(infos, expected_infos, map, worker) {
  assert_equals(infos.length, expected_infos.length,
                `request length for ${worker}`);
  for (var i = 0; i < infos.length; i++) {
    const key = expected_infos[i].resultingClientId;
    const url = expected_infos[i].url;
    const actual_id = infos[i].resultingClientId;
    const expected_id = map[key];
    // assert_equals(typeof(actual_id), 'string', `resultingClientId for ${url} should be a string`);
    if (expected_id) {
      assert_equals(requestInfos[0], expected_id,
                    `resultingClientId for ${url}`);
    } else {
      map[key] = actual_id;
    }
  }
}

// Creates an iframe and navigates to |url|, which is expected to start a chain
// of redirects.
// - |expected_last_url| is the expected window.location after the
//   navigation.
// - |expected_intercepted_urls| is the expected URLs that the service
//   workers were dispatched fetch events for. The format is:
//   [
//     [urls from workers[0]],
//     [urls from workers[1]],
//     [urls from cross-origin worker]
//   ]
function redirect_test(url,
                       expected_last_url,
                       expected_request_infos,
                       test_name) {
  promise_test(async t => {
    const message_promise = new Promise(resolve => {
        // A message with ID 'last_url' will be sent from the iframe.
        message_resolvers['last_url'] = resolve;
    });
    const frame = await with_iframe(url);
    t.add_cleanup(() => { frame.remove(); });

    // The "client" variant tests the Clients API using resultingClientId.
    if (document.location.search == '?client') {
      const map = {};
      // Request infos is an array like:
      // [
      //   [{url: url1, resultingClientId: id1}],
      //   [{url: url2, resultingClientId: id1}],
      //   [{url: url3: resultingClientId: id2}]
      // ]
      const requestInfos = await get_all_request_infos();
      check_resulting_client_ids(requestInfos[0], expected_request_infos[0], map, 'worker0');
      check_resulting_client_ids(requestInfos[1], expected_request_infos[1], map, 'worker1');
      check_resulting_client_ids(requestInfos[2], expected_request_infos[2], map, 'crossOriginWorker');

      // Now map is like:
      // {x: id1, b: id2, c: id3}
      // Ask each worker to try to resolve the ids to clients.
      // Only 'x' should resolve to a client.
      const client_infos  = await get_all_clients(map);

      // Client infos is an object like:
      // {
      //   worker0: {x: {found: true, id: id1, url: url1}, b: {found: false}},
      //   worker1: {x: {found: true, id: id1, url: url1}},
      //   crossOriginWorker: {x: {found: false}, b, {found: false}]
      // }
      // Only 'x' should be found, id should be id1, and url should be expected_last_url.
      // At least one 'x' should be present.
      let found = false;
      let expected_id = map['x'];
      found = check_clients(client_infos.worker0,
                            expected_id,
                            expected_last_url,
                            'worker0');
      found = check_clients(client_infos.worker1,
                            expected_id,
                            expected_last_url,
                            'worker1') || found;
      found = check_clients(client_infos.crossOriginWorker,
                            expected_id,
                            expected_last_url,
                            'crossOriginWorker') || found;
      assert_true(found, 'a client should have been found');
    }
    // The default variant tests the request interception chain and
    // resulting document.location.
    else {
      const expected_intercepted_urls = expected_request_infos.map(
          request_list => {
        return request_list.map(info => {
          return info.url;
        });
      });
      console.log(expected_intercepted_urls);
      console.log('check');
      await check_all_intercepted_urls(expected_intercepted_urls);
      console.log('last');
      const last_url = await message_promise;
      console.log('last2');
      assert_equals(last_url, expected_last_url, 'Last URL should match.');
    }
  }, test_name);
}

window.addEventListener('message', on_message, false);

function on_message(e) {
  if (e.origin != host_info['HTTPS_REMOTE_ORIGIN'] &&
      e.origin != host_info['HTTPS_ORIGIN'] ) {
    console.error('invalid origin: ' + e.origin);
    return;
  }
  var resolve = message_resolvers[e.data.id];
  delete message_resolvers[e.data.id];
  if (resolve) {
    resolve(e.data.result);
  }
}

function send_to_iframe(frame, message) {
  var message_id = next_message_id++;
  return new Promise(resolve => {
    message_resolvers[message_id] = resolve;
    frame.contentWindow.postMessage(
        {id: message_id, message},
        host_info['HTTPS_REMOTE_ORIGIN']);
  });
}

async function get_all_clients(map) {
  const client_infos = {};
  client_infos['worker0'] = await get_clients(workers[0], map);
  client_infos['worker1'] = await get_clients(workers[1], map);
  client_infos['crossOriginWorker'] =
      await send_to_iframe(other_origin_frame,
                           {command: 'get_clients', map: map});
  return client_infos;
}

function get_clients(worker, map) {
  return new Promise(resolve => {
    var channel = new MessageChannel();
    channel.port1.onmessage = (msg) => {
      resolve(msg.data.clients);
    };
    worker.postMessage({message: 'getClients', map, port: channel.port2},
                       [channel.port2]);
  });
}

function get_intercepted_urls(worker) {
  return new Promise(resolve => {
    var channel = new MessageChannel();
    channel.port1.onmessage = (msg) => {
      const urls = msg.data.requestInfos.map(info => {
        return info.url;
      });
      resolve(urls);
    };
    worker.postMessage({message: 'getRequestInfos', port: channel.port2}, [channel.port2]);
  });
}

function get_request_infos(worker) {
  return new Promise(resolve => {
    var channel = new MessageChannel();
    channel.port1.onmessage = (msg) => {
      resolve(msg.data.requestInfos);
    };
    worker.postMessage({message: 'getRequestInfos', port: channel.port2}, [channel.port2]);
  });
}

let url;

// Normal redirect (from out-scope to in-scope).
url = SCOPE1;
redirect_test(
    OUT_SCOPE + 'url=' + encodeURIComponent(url),
    url,
    [[{url, resultingClientId: 'x'}], [], []],
    'Normal redirect to same-origin scope.');


url = SCOPE1 + '#ref';
redirect_test(
    OUT_SCOPE + 'url=' + encodeURIComponent(SCOPE1) + '#ref',
    url,
    [[{url, resultingClientId: 'x'}], [], []],
    'Normal redirect to same-origin scope with a hash fragment.');

url = SCOPE1 + '#ref2';
redirect_test(
    OUT_SCOPE + 'url=' + encodeURIComponent(url) + '#ref',
    url,
    [[{url, resultingClientId: 'x'}], [], []],
    'Normal redirect to same-origin scope with different hash fragments.');

url = OTHER_ORIGIN_SCOPE;
redirect_test(
    OUT_SCOPE + 'url=' + encodeURIComponent(url),
    url,
    [[], [], [{url, resultingClientId: 'x'}]],
    'Normal redirect to other-origin scope.');

/*
// SW fallbacked redirect. SW doesn't handle the fetch request.
redirect_test(
    SCOPE1 + 'url=' + encodeURIComponent(OUT_SCOPE),
    OUT_SCOPE,
    [[SCOPE1 + 'url=' + encodeURIComponent(OUT_SCOPE)], [], []],
    'SW-fallbacked redirect to same-origin out-scope.');
redirect_test(
    SCOPE1 + 'url=' + encodeURIComponent(SCOPE1),
    SCOPE1,
    [[SCOPE1 + 'url=' + encodeURIComponent(SCOPE1), SCOPE1], [], []],
    'SW-fallbacked redirect to same-origin same-scope.');
redirect_test(
    SCOPE1 + 'url=' + encodeURIComponent(SCOPE1) + '#ref',
    SCOPE1 + '#ref',
    [
      [SCOPE1 + 'url=' + encodeURIComponent(SCOPE1) + '#ref', SCOPE1 + '#ref'],
      [],
      []
    ],
    'SW-fallbacked redirect to same-origin same-scope with a hash fragment.');
redirect_test(
    SCOPE1 + 'url=' + encodeURIComponent(SCOPE1 + '#ref2') + '#ref',
    SCOPE1 + '#ref2',
    [
      [
        SCOPE1 + 'url=' + encodeURIComponent(SCOPE1 + '#ref2') + '#ref',
        SCOPE1 + '#ref2'
      ],
      [],
      []
    ],
    'SW-fallbacked redirect to same-origin same-scope with different hash ' +
    'fragments.');
redirect_test(
    SCOPE1 + 'url=' + encodeURIComponent(SCOPE2),
    SCOPE2,
    [[SCOPE1 + 'url=' + encodeURIComponent(SCOPE2)], [SCOPE2], []],
    'SW-fallbacked redirect to same-origin other-scope.');
redirect_test(
    SCOPE1 + 'url=' + encodeURIComponent(OTHER_ORIGIN_OUT_SCOPE),
    OTHER_ORIGIN_OUT_SCOPE,
    [[SCOPE1 + 'url=' + encodeURIComponent(OTHER_ORIGIN_OUT_SCOPE)], [], []],
    'SW-fallbacked redirect to other-origin out-scope.');
redirect_test(
    SCOPE1 + 'url=' + encodeURIComponent(OTHER_ORIGIN_SCOPE),
    OTHER_ORIGIN_SCOPE,
    [
      [SCOPE1 + 'url=' + encodeURIComponent(OTHER_ORIGIN_SCOPE)],
      [],
      [OTHER_ORIGIN_SCOPE]
    ],
    'SW-fallbacked redirect to other-origin in-scope.');

// SW generated redirect.
// SW: event.respondWith(Response.redirect(params['url']));
redirect_test(
    SCOPE1 + 'sw=gen&url=' + encodeURIComponent(OUT_SCOPE),
    OUT_SCOPE,
    [[SCOPE1 + 'sw=gen&url=' + encodeURIComponent(OUT_SCOPE)], [], []],
    'SW-generated redirect to same-origin out-scope.');
redirect_test(
    SCOPE1 + 'sw=gen&url=' + encodeURIComponent(OUT_SCOPE) + '#ref',
    OUT_SCOPE + '#ref',
    [
      [SCOPE1 + 'sw=gen&url=' + encodeURIComponent(OUT_SCOPE) + '#ref'],
      [],
      []
    ],
    'SW-generated redirect to same-origin out-scope with a hash fragment.');
redirect_test(
    SCOPE1 + 'sw=gen&url=' + encodeURIComponent(OUT_SCOPE + '#ref2') + '#ref',
    OUT_SCOPE + '#ref2',
    [
      [SCOPE1 + 'sw=gen&url=' + encodeURIComponent(OUT_SCOPE + '#ref2') + '#ref'],
      [],
      []
    ],
    'SW-generated redirect to same-origin out-scope with different hash ' +
    'fragments.');
redirect_test(
    SCOPE1 + 'sw=gen&url=' + encodeURIComponent(SCOPE1),
    SCOPE1,
    [[SCOPE1 + 'sw=gen&url=' + encodeURIComponent(SCOPE1), SCOPE1], [], []],
    'SW-generated redirect to same-origin same-scope.');
redirect_test(
    SCOPE1 + 'sw=gen&url=' + encodeURIComponent(SCOPE2),
    SCOPE2,
    [[SCOPE1 + 'sw=gen&url=' + encodeURIComponent(SCOPE2)], [SCOPE2], []],
    'SW-generated redirect to same-origin other-scope.');
redirect_test(
    SCOPE1 + 'sw=gen&url=' + encodeURIComponent(OTHER_ORIGIN_OUT_SCOPE),
    OTHER_ORIGIN_OUT_SCOPE,
    [
      [SCOPE1 + 'sw=gen&url=' + encodeURIComponent(OTHER_ORIGIN_OUT_SCOPE)],
      [],
      []
    ],
    'SW-generated redirect to other-origin out-scope.');
redirect_test(
    SCOPE1 + 'sw=gen&url=' + encodeURIComponent(OTHER_ORIGIN_SCOPE),
    OTHER_ORIGIN_SCOPE,
    [
      [SCOPE1 + 'sw=gen&url=' + encodeURIComponent(OTHER_ORIGIN_SCOPE)],
      [],
      [OTHER_ORIGIN_SCOPE]
    ],
    'SW-generated redirect to other-origin in-scope.');

// SW fetched redirect.
// SW: event.respondWith(fetch(event.request));
redirect_test(
    SCOPE1 + 'sw=fetch&url=' + encodeURIComponent(OUT_SCOPE),
    OUT_SCOPE,
    [[SCOPE1 + 'sw=fetch&url=' + encodeURIComponent(OUT_SCOPE)],
    [],
    []],
    'SW-fetched redirect to same-origin out-scope.');
redirect_test(
    SCOPE1 + 'sw=fetch&url=' + encodeURIComponent(SCOPE1),
    SCOPE1,
    [[SCOPE1 + 'sw=fetch&url=' + encodeURIComponent(SCOPE1), SCOPE1], [], []],
    'SW-fetched redirect to same-origin same-scope.');
redirect_test(
    SCOPE1 + 'sw=fetch&url=' + encodeURIComponent(SCOPE2),
    SCOPE2,
    [
      [SCOPE1 + 'sw=fetch&url=' + encodeURIComponent(SCOPE2)],
      [SCOPE2],
      []
    ],
    'SW-fetched redirect to same-origin other-scope.');
redirect_test(
    SCOPE1 + 'sw=fetch&url=' + encodeURIComponent(OTHER_ORIGIN_OUT_SCOPE),
    OTHER_ORIGIN_OUT_SCOPE,
    [
      [SCOPE1 + 'sw=fetch&url=' + encodeURIComponent(OTHER_ORIGIN_OUT_SCOPE)],
      [],
      []
    ], 'SW-fetched redirect to other-origin out-scope.');
redirect_test(
    SCOPE1 + 'sw=fetch&url=' + encodeURIComponent(OTHER_ORIGIN_SCOPE),
    OTHER_ORIGIN_SCOPE,
    [
      [SCOPE1 + 'sw=fetch&url=' + encodeURIComponent(OTHER_ORIGIN_SCOPE)],
      [],
      [OTHER_ORIGIN_SCOPE]
    ],
    'SW-fetched redirect to other-origin in-scope.');

// Opaque redirect.
// SW: event.respondWith(fetch(
//         new Request(event.request.url, {redirect: 'manual'})));
redirect_test(
    SCOPE1 + 'sw=manual&url=' + encodeURIComponent(OUT_SCOPE),
    OUT_SCOPE,
    [[SCOPE1 + 'sw=manual&url=' + encodeURIComponent(OUT_SCOPE)], [], []],
    'Redirect to same-origin out-scope with opaque redirect response.');
redirect_test(
    SCOPE1 + 'sw=manual&url=' + encodeURIComponent(SCOPE1),
    SCOPE1,
    [[SCOPE1 + 'sw=manual&url=' + encodeURIComponent(SCOPE1), SCOPE1], [], []],
    'Redirect to same-origin same-scope with opaque redirect response.');
redirect_test(
    SCOPE1 + 'sw=manual&url=' + encodeURIComponent(SCOPE2),
    SCOPE2,
    [[SCOPE1 + 'sw=manual&url=' + encodeURIComponent(SCOPE2)], [SCOPE2], []],
    'Redirect to same-origin other-scope with opaque redirect response.');
redirect_test(
    SCOPE1 + 'sw=manual&url=' +
    encodeURIComponent(OTHER_ORIGIN_OUT_SCOPE),
    OTHER_ORIGIN_OUT_SCOPE,
    [
      [SCOPE1 + 'sw=manual&url=' + encodeURIComponent(OTHER_ORIGIN_OUT_SCOPE)],
      [],
      []
    ],
    'Redirect to other-origin out-scope with opaque redirect response.');
redirect_test(
    SCOPE1 + 'sw=manual&url=' + encodeURIComponent(OTHER_ORIGIN_SCOPE),
    OTHER_ORIGIN_SCOPE,
    [
      [SCOPE1 + 'sw=manual&url=' + encodeURIComponent(OTHER_ORIGIN_SCOPE)],
      [],
      [OTHER_ORIGIN_SCOPE]
    ],
    'Redirect to other-origin in-scope with opaque redirect response.');
redirect_test(
    SCOPE1 + 'sw=manual&noLocationRedirect',
    SCOPE1 + 'sw=manual&noLocationRedirect',
    [[SCOPE1 + 'sw=manual&noLocationRedirect'], [], []],
    'No location redirect response.');

// Opaque redirect passed through Cache.
// SW responds with an opaque redirectresponse from the Cache API.
redirect_test(
    SCOPE1 + 'sw=manualThroughCache&url=' + encodeURIComponent(OUT_SCOPE),
    OUT_SCOPE,
    [
      [SCOPE1 + 'sw=manualThroughCache&url=' + encodeURIComponent(OUT_SCOPE)],
      [],
      []
    ],
    'Redirect to same-origin out-scope with opaque redirect response which ' +
    'is passed through Cache.');
redirect_test(
    SCOPE1 + 'sw=manualThroughCache&url=' + encodeURIComponent(SCOPE1),
    SCOPE1,
    [
      [
        SCOPE1 + 'sw=manualThroughCache&url=' + encodeURIComponent(SCOPE1),
        SCOPE1
      ],
      [],
      []
    ],
    'Redirect to same-origin same-scope with opaque redirect response which ' +
    'is passed through Cache.');
redirect_test(
    SCOPE1 + 'sw=manualThroughCache&url=' + encodeURIComponent(SCOPE2),
    SCOPE2,
    [
      [SCOPE1 + 'sw=manualThroughCache&url=' + encodeURIComponent(SCOPE2)],
      [SCOPE2],
      []
    ],
    'Redirect to same-origin other-scope with opaque redirect response which ' +
    'is passed through Cache.');
redirect_test(
    SCOPE1 + 'sw=manualThroughCache&url=' +
        encodeURIComponent(OTHER_ORIGIN_OUT_SCOPE),
    OTHER_ORIGIN_OUT_SCOPE,
    [
      [SCOPE1 + 'sw=manualThroughCache&url=' +
          encodeURIComponent(OTHER_ORIGIN_OUT_SCOPE)
      ],
      [],
      []
    ],
    'Redirect to other-origin out-scope with opaque redirect response which ' +
    'is passed through Cache.');
redirect_test(
    SCOPE1 + 'sw=manualThroughCache&url=' +
        encodeURIComponent(OTHER_ORIGIN_SCOPE),
    OTHER_ORIGIN_SCOPE,
    [
      [SCOPE1 + 'sw=manualThroughCache&url=' +
          encodeURIComponent(OTHER_ORIGIN_SCOPE)],
       [],
       [OTHER_ORIGIN_SCOPE],
    ],
    'Redirect to other-origin in-scope with opaque redirect response which ' +
    'is passed through Cache.');
redirect_test(
    SCOPE1 + 'sw=manualThroughCache&noLocationRedirect',
    SCOPE1 + 'sw=manualThroughCache&noLocationRedirect',
    [[SCOPE1 + 'sw=manualThroughCache&noLocationRedirect'], [], []],
    'No location redirect response via Cache.');
*/
// Clean up the test environment. This promise_test() needs to be the last one.
promise_test(async t => {
  registrations.forEach(async registration => {
    if (registration)
      await registration.unregister();
  });
  await send_to_iframe(other_origin_frame, {command: 'unregister'});
  other_origin_frame.remove();
}, 'clean up global state');
</script>
</body>
